pyproject.toml
[project]
name = "advanced-fastapi"
version = "0.1.0"
description = "Advanced FastAPI Application"
readme = "README.md"
authors = [
    { name = "BapanBigData", email = "bapanmldl7892@gmail.com" }
]
requires-python = ">=3.11,<3.12"
dependencies = [
    # use redis>=5 which provides redis.asyncio; remove aioredis
    "redis>=5.0.0",
    "alembic>=1.16.5",
    "asyncmy>=0.2.10",
    "asyncpg>=0.30.0",
    "bcrypt>=4.0.0,<4.1.0",
    "cryptography>=45.0.6",
    "fastapi>=0.116.1",
    "passlib[bcrypt]==1.7.4",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.24",
    "uvicorn[standard]>=0.35.0",
    "fastapi-mail>=1.5.0",
    "itsdangerous>=2.2.0",
    "celery>=5.5.3",
    "asgiref>=3.9.2",
    "flower>=2.0.1",
    "pytest>=8.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.10",
    "types-requests>=2.32.4.20250809",
]
